+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Remove multiple markers (#596)
 * Add more lua functions (savestates, ff)
 * Implement ppoll
+* Implement snd_async_handler_get_callback_private
 
 ### Changed